Window function improvements: decimal RANGE bounds, NULL ordering, frame validation
- Support decimal RANGE frame bounds (e.g., 0.5D->_range(2.5))
  - Updated FrameBound records to use double instead of int
  - Updated parsers to use doubleValue() for decimal support
  - SQL generator formats clean numbers (1 not 1.0)

- Add NULLS FIRST/LAST to window ORDER BY for Pure semantics
  - DESC uses NULLS FIRST, ASC uses NULLS LAST
  - Matches Pure's expected NULL ordering behavior

- Validate window frame boundaries at construction time
  - Reject invalid frames where lower bound > upper bound
  - e.g., 4->_range(2) throws IllegalArgumentException

- Fix PCT error message passthrough (remove wrapper prefix)

- Add integration tests for all new features

+++ engine/src/main/java/org/finos/legend/pure/dsl/PureCompiler.java
@@ -2130,14 +2130,15 @@ public final class PureCompiler {
     }
 
     /**
-     * Parses a frame bound: unbounded(), 0 (current row), or integer
-     * (preceding/following)
+     * Parses a frame bound: unbounded(), 0 (current row), or number
+     * (preceding/following).
+     * Supports decimal bounds for RANGE frames (e.g., 0.5D->_range(2.5)).
      */
     private WindowExpression.FrameBound parseFrameBound(PureExpression expr) {
         if (expr instanceof FunctionCall fc && "unbounded".equals(fc.functionName())) {
             return WindowExpression.FrameBound.unbounded();
         } else if (expr instanceof LiteralExpr lit && lit.value() instanceof Number num) {
-            int value = num.intValue();
+            double value = num.doubleValue();
             if (value == 0) {
                 return WindowExpression.FrameBound.currentRow();
             } else if (value < 0) {
@@ -2146,9 +2147,9 @@ public final class PureCompiler {
                 return WindowExpression.FrameBound.following(value);
             }
         } else if (expr instanceof UnaryExpression unary && "-".equals(unary.operator())) {
-            // Handle negative literals like -1, -2
+            // Handle negative literals like -1, -2, -0.5
             if (unary.operand() instanceof LiteralExpr lit && lit.value() instanceof Number num) {
-                int value = num.intValue();
+                double value = num.doubleValue();
                 return WindowExpression.FrameBound.preceding(value); // Already positive, represents "X PRECEDING"
             }
         }
